Got resizing working with base bubble; ClassicSpeechBubble now inherits from BaseDialogBubble
